Hello Aleo world
"One of Us" melds the classic "innocent vs traitor" gameplay with state-of-the-art zero knowledge on-chain mechanics. Drawing inspiration from beloved games like Mafia, Avalon, and Among Us, it's not just a game – it’s a journey into the depths of strategy, suspense, and social interaction whose goal is to grow and strengthen the Aleo community.